+++ sdwan_policy_objects.tf
@@ -4,13 +4,14 @@ resource "sdwan_policy_object_app_probe_class" "policy_object_app_probe_class" {
   description        = null # not supported in the UI
   feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
   entries = [{
-    forwarding_class = each.value.forwarding_class
+    forwarding_class_id = sdwan_policy_object_class_map.policy_object_class_map[each.value.forwarding_class].id
     map = [for m in try(each.value.mappings, []) : {
       color = m.color
       dscp  = try(m.dscp, null)
     }]
   }]
 }
+
 resource "sdwan_policy_object_application_list" "policy_object_application_list" {
   for_each           = { for p in try(local.feature_profiles.policy_object_profile.application_lists, {}) : p.name => p }
   name               = each.value.name
@@ -250,6 +251,32 @@ resource "sdwan_policy_object_security_url_block_list" "policy_object_security_u
   }]
 }
 
+resource "sdwan_policy_object_security_zone" "policy_object_security_zone" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.security_zones, {}) : p.name => p }
+  name               = each.value.name
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for e in concat([for v in try(each.value.vpns, []) : { "vpn" : v }], [for intf in try(each.value.interfaces, []) : { "interface" : intf }]) : {
+    vpn       = try(e.vpn, null)
+    interface = try(e.interface, null)
+  }]
+}
+
+resource "sdwan_policy_object_unified_url_filtering" "policy_object_unified_url_filtering" {
+  for_each              = { for p in try(local.feature_profiles.policy_object_profile.security_url_filtering_profiles, {}) : p.name => p }
+  name                  = each.value.name
+  feature_profile_id    = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  alerts                = try(each.value.alerts, null)
+  block_page_action     = each.value.block_page_action
+  block_page_contents   = each.value.block_page_action == "text" ? try(each.value.block_page_content_body, null) != null ? "Access to the requested page has been denied. ${each.value.block_page_content_body}" : "Access to the requested page has been denied. ${local.defaults.sdwan.feature_profiles.policy_object_profile.security_url_filtering_profiles.block_page_content_body}" : null
+  enable_alerts         = each.value.enable_alerts
+  redirect_url          = try(each.value.redirect_url, null)
+  url_allow_list_id     = try(sdwan_policy_object_security_url_allow_list.policy_object_security_url_allow_list[each.value.url_allow_list].id, null)
+  url_block_list_id     = try(sdwan_policy_object_security_url_block_list.policy_object_security_url_block_list[each.value.url_block_list].id, null)
+  web_categories        = each.value.web_categories
+  web_categories_action = each.value.web_categories_action
+  web_reputation        = each.value.web_reputation
+}
+
 resource "sdwan_policy_object_sla_class_list" "policy_object_sla_class_list" {
   for_each           = { for p in try(local.feature_profiles.policy_object_profile.sla_classes, {}) : p.name => p }
   name               = each.value.name
@@ -288,6 +315,18 @@ resource "sdwan_policy_object_tloc_list" "policy_object_tloc_list" {
     tloc_ip       = e.tloc_ip
     preference    = try(e.preference, null)
   }]
+}
+
+resource "sdwan_policy_object_unified_advanced_inspection_profile" "policy_object_unified_advanced_inspection_profile" {
+  for_each                            = { for p in try(local.feature_profiles.policy_object_profile.security_advanced_inspection_profiles, {}) : p.name => p }
+  name                                = each.value.name
+  description                         = try(each.value.description, null)
+  feature_profile_id                  = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  tls_decryption_action               = lookup({ "never_decrypt" = "neverDecrypt", "skip_decrypt" = "skipDecrypt" }, each.value.tls_action, each.value.tls_action)
+  intrusion_prevention_list_id        = try(sdwan_policy_object_unified_intrusion_prevention.policy_object_unified_intrusion_prevention[each.value.intrusion_prevention].id, null)
+  url_filtering_list_id               = try(sdwan_policy_object_unified_url_filtering.policy_object_unified_url_filtering[each.value.url_filtering].id, null)
+  advanced_malware_protection_list_id = try(sdwan_policy_object_unified_advanced_malware_protection.policy_object_unified_advanced_malware_protection[each.value.advanced_malware_protection].id, null)
+  tls_ssl_profile_list_id             = null # not supported
 }
 
 resource "sdwan_policy_object_unified_advanced_malware_protection" "policy_object_unified_advanced_malware_protection" {